refactor: extract shared LLM retry/backoff helper; remove dead ChatService field (#1702)
Part of #1702's staged consolidation of the server's five parallel LLM
invocation paths. This lands the two lowest-risk, purely additive steps
from the issue's implementation plan:

- Extract WorkflowLLMHelper's transient-error retry/backoff logic
  (isTransientHttpStatus, isTransientLlmError, parseRetryAfterMs,
  computeRetryDelayMs, runWithRetries) into a new shared module,
  server/services/llm/retryWithBackoff.js, so other LLM invocation paths
  can reuse it instead of having no retry at all. WorkflowLLMHelper
  re-exports the same names for backward compatibility.
- Apply the shared retry wrapper to simpleCompletion (server/utils.js),
  which previously had zero retry — a single transient 503/429 failed
  magic-prompt/translate/model-test/agent-tool calls outright.
- Delete PromptNodeExecutor's unused ChatService field/import (confirmed
  dead: zero reads anywhere in the file) and its stale docstring claim.

The remaining, higher-risk steps from the issue's plan (canonical SSE
parser, merging the two tool-calling loops, normalizing usage accounting
across all five paths) are left as follow-up, per the plan's own staged
sequencing recommendation.
